soc/apps/api/requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
websockets==12.0
alembic==1.13.1